Performance enhancement
Problem with previous implementation was that if scrolling to the bottom of the list, the rendering would slower and swlower depending on the list's length. The items would go down from top to bottom rendering every single step until it reaches the current scroll position. For like 250 items of 110px height it would take like 6-7 seconds to render the list.

With my implementation items are now placed directly at the proper position with the appropriate data.
They're populated only when necessary.